requirements.txt
streamlit==1.32.0
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
nltk==3.8.1
python-dotenv==1.0.1
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1 